PERF: Only transform functional styles
While testing victory#141, I did some profiling and found that
`evaluateStyles` was the most time-consuming function. It's currently
transforming styles even if it would be a no-op.

This only transforms style objects that contain functional styles. This
increased FPS by ~30% for me. The most time consuming functions are now
all from Radium, as I expected. We'll need to focus our efforts
optimizing Radium (or avoiding it) if we'd like to make Victory much
faster.

/cc @boygirl @bclinkinbeard